refactor(agent): decompose ReviewerAgent into Review collaborators
ReviewerAgent had grown to 715 lines doing seven jobs (mode resolution,
verdict application, outcome/coverage recording, cache adaptation, and
the sequential/concurrent/batch strategies). Split per the
single-responsibility rule into cohesive Application\Agent\Review
collaborators:

- VerdictApplier — applies one verdict payload (acceptance, severity
  elevation, type correction) to a finding.
- BatchVerdictApplier — matches batch verdicts to findings by id and
  degrades failed batches to rejected/errored.
- ReviewOutcomeRecorder — turns verdicts, raw LLM responses, and
  failures into the reviewed finding plus its coverage entry.
- ReviewerVerdictCache — adapts the optional ReviewerCacheInterface
  (bypass/missing cache degrade to a miss).
- CodeContextResolver — resolves a finding's source content.
- Sequential/Structured/Concurrent/ConcurrentStructured/Batch
  ReviewAnalyzer — the five analysis strategies.

ReviewerAgent (now ~200 lines, mostly the constructor) builds these from
its existing injected dependencies, so the public constructor signature
and the DI wiring are unchanged and every review()-level test exercises
the collaborators unchanged. Behaviour is identical — full suite green.
